Phase A: split /p/<shortId> into /t/<threadShortId>?p=<pollShortId>
The single /p/<shortId> route did triple duty as thread anchor, specific-poll
deep-link, and "id of any kind" resolver. Split into:

  /t/<threadShortId>            view a thread (no auto-expand, scroll to bottom)
  /t/<threadShortId>?p=<id>     view a thread with that poll expanded + scrolled to
  /p/<shortId>                  legacy redirect: resolves and 302s into /t/...

Removes the brittle bits the overload required: the ?thread=1 query flag
(emitted but no longer consumed), the suppressExpand "user has responded to
every question" heuristic, and the path-param resolution cascade (poll
short_id / poll uuid / question uuid). The new /t/ route's path is
unambiguously a thread root short_id; the optional ?p= names a specific poll
to expand. URL sync on expand now swaps ?p= instead of replacing the whole path.

Phase B (server-side thread membership with browser_id cookie + materialized
threads table) and Phase C (join-time visibility) are deferred. See
docs/thread-routing-redesign.md for the full plan.
