+++ app/poll/page.tsx
@@ -14,8 +14,10 @@ function QuestionRedirect() {
 
   useEffect(() => {
     if (id) {
-      // Redirect to new URL structure
-      router.replace(`/p/${id}`);
+      // Hand off to /t/?id=<id> which resolves the question id to its thread
+      // root and redirects to `/t/<root>?p=<pollShortId>`. Preserves the
+      // legacy /poll?id=<question-uuid> deep-link form.
+      router.replace(`/t/?id=${encodeURIComponent(id)}`);
     } else {
       // If no id, redirect to home
       router.replace('/');
